parse_rprof {profvis}	R Documentation

Parse Rprof output file for use with profvis

Description

Parse Rprof output file for use with profvis

Usage

parse_rprof(path = "Rprof.out", expr_source = NULL)

Arguments

path
Path to the Rprof output file.

expr_source
If any source refs in the profiling output have an empty filename, that means they refer to code executed at the R console. This code can be captured and passed (as a string) as the expr_source argument.

[Package profvis version 0.3.6 Index]