pause {profvis}	R Documentation

Pause an R process

Description

This function pauses an R process for some amount of time. It differs from Sys.sleep in that time spent in pause will show up in profiler data. Another difference is that pause uses up 100 whereas Sys.sleep does not.

Usage

pause(seconds)

Arguments

seconds
Number of seconds to pause.

Examples

# Wait for 0.5 seconds
pause(0.5)

[Package profvis version 0.3.6 Index]